+++ app/javascript/vue/tasks/sources/new_source/components/rightSection.vue
@@ -7,7 +7,7 @@
         ref="documents"
         class="panel"/>
       <soft-validation
-        v-if="validations"
+        v-if="areValidations"
         class="margin-medium-top soft-validation-panel"
         :validations="validations"/>
       <matches-component ref="matches"/>
@@ -30,6 +30,9 @@ export default {
   computed: {
     validations () {
       return this.$store.getters[GetterNames.GetSoftValidation]
+    },
+    areValidations () {
+      return this.validations?.sources?.list.length
     }
   },
   data () {