+++ app/javascript/vue/tasks/sources/new_source/store/actions/actionNames.js
@@ -2,6 +2,7 @@ const ActionNames = {
   CloneSource: 'cloneSource',
   ConvertToBibtex: 'convertToBibtex',
   LoadSource: 'loadSource',
+  LoadSoftValidations: 'loadSoftValidations',
   ResetSource: 'resetSource',
   SaveDocument: 'saveDocument',
   SaveDocumentation: 'saveDocumentation',